Move PerfType to a separate module
