+++ src/leaderboard/page/html_generator.py
@@ -7,7 +7,7 @@ from jinja2 import Environment, FileSystemLoader
 from src.leaderboard.chrono import date_formatter, duration_formatter
 from src.leaderboard.chrono.time_provider import TimeProvider
 from src.leaderboard.data.leaderboard_row import LeaderboardRow
-from src.leaderboard.li.bot_user import PerfType
+from src.leaderboard.li.pert_type import PerfType
 
 
 @dataclasses.dataclass(frozen=True)
